% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_etal.R
\name{authors_etal}
\alias{authors_etal}
\title{Format author strings}
\usage{
authors_etal(x, authors = 3, split = ", *")
}
\arguments{
\item{x}{A vector of author names}

\item{authors}{Number of authors to display before adding et al.}

\item{split}{Character used for splitting author strings, default comma}
}
\value{
a vector
}
\description{
Format delimited lists of authors and replace others with et al.
}
\details{
Replaces 2 or more authors with et al.
}
\examples{
authors_etal("Kawasaki S, Mizuguchi K,Sato M, Kono T, Shimizu H.")
authors_etal("Kawasaki S, Mizuguchi K,Sato M, Kono T, Shimizu H.", 2)
}
\author{
Chris Stubben
}
