% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_format.R
\name{bib_format}
\alias{bib_format}
\title{Format a bibliography}
\usage{
bib_format(x, authors = 3, issue = TRUE, links = FALSE,
  cited = FALSE, pmid = FALSE, number = FALSE)
}
\arguments{
\item{x}{epmc_search results}

\item{authors}{Number of authors to display before adding et al.}

\item{issue}{Include issue number with citation}

\item{links}{Add Markdown links to Journal, Cited by counts and PubMed ID, default FALSE}

\item{cited}{Include Cited By:<count>}

\item{pmid}{Include PubMed:<ID>}

\item{number}{Number references in bibliography}
}
\value{
a vector
}
\description{
Format Europe PMC search results into a bibliography
}
\note{
Currently, the references are formatted using author, year, title and journal
}
\examples{
data(yp)
bib_format(yp[1:5,])
cat(strwrap(bib_format(yp[1:5,], number=TRUE, pmid=TRUE), exdent=3), sep="\\n")
}
\seealso{
\code{\link{strwrap}} for formatting
}
\author{
Chris Stubben
}
