% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mesh.R
\name{format_mesh}
\alias{format_mesh}
\title{Format MeSH terms}
\usage{
format_mesh(x)
}
\arguments{
\item{x}{a vector of MeSH terms}
}
\value{
a tibble
}
\description{
Format a summary table of semi-colon delimited MeSH terms
(with major topics marked by *)
}
\examples{
data(yp)
format_mesh(yp)
}
\author{
Chris Stubben
}
