% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlm}
\alias{nlm}
\title{Journals currently or previously indexed in MEDLINE}
\format{A data frame with 15001 observations on the following 7 variables.
  \describe{
    \item{ta}{ title abbreviation = MedlineTA tag}
    \item{title}{ journal title = TitleMain/Title tag}
    \item{country}{ Country tag}
    \item{language}{ primary language = Language[@LangType="Primary"] tag }
    \item{year}{ PublicationFirstYear tag }
    \item{mesh}{ MeSH terms including descriptor, qualifier and major topic marked with * }
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/nlmcatalog}
}
\usage{
nlm
}
\description{
Titles, abbreviations, MeSH terms and other details of journals currently or
previously indexed in the NLM Catalog at NCBI
}
\details{
XML search results from the NLM Catalog at NCBI using "reportedmedline" on April 18, 2016
}
\examples{
data(nlm)
filter(nlm, grepl("Heart Diseases", mesh) )
sort( table(unlist(strsplit(nlm$mesh, "; "))), decreasing=TRUE)[1:10]

}
\keyword{datasets}
