% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_core.R
\name{search_core}
\alias{search_core}
\title{Search Europe PMC core and get article citations, abstract and MeSH terms}
\usage{
search_core(query, page, limit = 1000)
}
\arguments{
\item{query}{a valid query string}

\item{page}{pagination options, default 1}

\item{limit}{number of articles, default 1000}
}
\value{
A data.frame with search results
}
\description{
Search the Europe PMC Rest Service and parse the core XML resulttype
}
\note{
Search fields are listed at \url{https://europepmc.org/Help#fieldsearch}
}
\examples{
search_core("Yersinia pestis hmsB")
}
\author{
Chris Stubben
}
\references{
\url{http://europepmc.org/RestfulWebService#search}
}

