% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_ts.R
\name{year_ts}
\alias{year_ts}
\title{Publication year time series}
\usage{
year_ts(x, start, end, sum = FALSE)
}
\arguments{
\item{x}{a vector of years, or optionally epmc_search results}

\item{start}{year of first observation, default minimum year in series}

\item{end}{year of last observation, default today's year}

\item{sum}{return cumulative sum}
}
\value{
a time-series object
}
\description{
Create a time-series object from publication years
}
\examples{
data(yp)
y <- year_ts(yp)
plot(y, xlab="Year published", ylab="Articles per year", las=1,
    main="Publications with Yersinia pestis virulence in title")
}
\author{
Chris Stubben
}
\seealso{
\code{\link{ts}}
}

