% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yp}
\alias{yp}
\title{Yersinia pestis virulence publications}
\format{A data frame with 160 observations on the following 19 variables.
 \describe{
   \item{pmid}{PubMed ID = pmid tag}
   \item{pmcid}{PubMed Central ID = pmcid tag}
   \item{src}{ source tag }
   \item{doi}{ DOI tag }
   \item{authors}{ authorString tag }
   \item{year}{ pubYear tag }
   \item{title}{ article title tag }
   \item{journal}{ full journal name = journal/title tag }
   \item{nlm_ta}{ journal abbreviation = medlineAbbreviation tag}
   \item{volume}{ journalInfo/volume tag }
   \item{issue}{ journalInfo/issue tag }
   \item{pages}{ pageInfo tag }
   \item{citedby}{ citedByCount tag }
   \item{published}{ firstPublicationDate tag }
   \item{language}{language tag}
   \item{abstract}{abstractText tag}
   \item{mesh}{ MeSH terms including descriptor, qualifier and major topic*}
   \item{keywords }{keyword tags}
   \item{chemicals }{chemical/name tags}
 }}
\source{
Europe PMC search
}
\usage{
yp
}
\description{
Europe PMC core search results with Yersinia pestis virulence in title
}
\examples{
data(yp)
\dontrun{ yp <- search_core("title:(Yersinia pestis virulence) src:MED")}
t(yp[7,])
bib_format(yp[1:7,])
data.frame(n=sort(table(unlist(strsplit(yp$pubType, "; "))), decreasing=TRUE))
subset(yp, grepl("retracted", pubType))
data.frame(n=sort(table(unlist(strsplit(yp$authorString, ", "))), decreasing=TRUE)[1:15])
}
\keyword{datasets}
