% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal_cite.R
\name{journal_cite}
\alias{journal_cite}
\title{Format journal citations}
\usage{
journal_cite(x, n, issue = TRUE)
}
\arguments{
\item{x}{a data.frame with output from \code{epmc_search}}

\item{n}{a vector with 4 columns positions for journal, volume, issue and page,
will match \code{epmc_search} output by default}

\item{issue}{include issue with citation, default TRUE}
}
\value{
a vector
}
\description{
Format journal title, volume, issue and pages into single string
}
\details{
Formats citations with or without the issue number like MBio 5(1):13  or MBio 5:13
}
\examples{
data(yp)
journal_cite(yp[1:5,])
journal_cite(yp[1:5,], issue=FALSE)
}
\author{
Chris Stubben
}

